requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyJWT
werkzeug
email-validator
pymysql
python-dotenv
